+++ gazebo/util/IntrospectionManager.hh
@@ -84,10 +84,15 @@ namespace gazebo
       /// through all the topics. The message received in the update will
       /// contain the name and latest values of all the items specified
       /// in the filter.
-      /// If there are changes in the items list since the last update, a new
-      /// message is published under the topic
+      /// If there are changes in the items list since the last update,
+      /// a new message is published under the topic
       /// "/introspection/<manager_id>/items_update".
       public: void Update();
+
+      /// \brief If there are changes in the items list since the last update,
+      /// a new message is published under the topic
+      /// "/introspection/<manager_id>/items_update".
+      public: void NotifyUpdates();
 
       /// \brief Constructor.
       private: IntrospectionManager();